made queries directory to mirror models directory. 1 file per db table
