+++ Webapp/sources/services/queries/__init__.py
@@ -1,0 +1,9 @@
+from . import (
+    compound,
+    compound_data_error_report,
+    reaction,
+    user,
+    workbook,
+    workgroup,
+    workgroup_request,
+)
