+++ Webapp/sources/services/queries/compound.py
@@ -1,0 +1,12 @@
+from sources import models
+from sources.extensions import db
+
+
+def get_compound_count() -> int:
+    """
+    Gets the number of compounds in the database
+
+    Returns:
+        The number of compounds in the database
+    """
+    return db.session.query(models.Compound).count()
