+++ Webapp/sources/services/queries/user.py
@@ -1,0 +1,14 @@
+from typing import List
+
+from sources import models
+from sources.extensions import db
+
+
+def get_all_users() -> List[models.User]:
+    """
+    Gets a list of all users in the database
+
+    Returns:
+         List of all users
+    """
+    return db.session.query(models.User).all()
