+++ Webapp/sources/services/queries/workgroup_request.py
@@ -1,0 +1,18 @@
+from typing import List
+
+from sources import models
+from sources.extensions import db
+
+
+def get_new_workgroup_requests() -> List[models.WorkGroupRequest]:
+    """
+    Gets a list of all workgroup requests in the database
+
+    Returns:
+         List of all workgroup requests
+    """
+    return (
+        db.session.query(models.WorkGroupRequest)
+        .filter(models.WorkGroupRequest.status == "active")
+        .all()
+    )
